Score working properly
Changed things to do that
