+++ core/src/org/grup7/deheroes/MyGdxGame.java
@@ -1,20 +1,32 @@
 package org.grup7.deheroes;
 
 import com.badlogic.gdx.Game;
+import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 
-import org.grup7.deheroes.helpers.AssetManager;
-import org.grup7.deheroes.screens.MenuScreen;
+import org.grup7.deheroes.Helpers.AssetManager;
+import org.grup7.deheroes.Screens.MenuScreen;
+
 
 public class MyGdxGame extends Game {
+    public SpriteBatch batch;
+
     @Override
     public void create() {
-        AssetManager.load();
-        setScreen(new MenuScreen(this));
-    }
+      batch = new SpriteBatch();
+      AssetManager.load();
+      setScreen(new MenuScreen(this));
 
+
+    }
+    @Override
+    public void render(){
+        super.render();
+    }
     @Override
     public void dispose() {
         super.dispose();
         AssetManager.dispose();
     }
+
+
 }
